final fixes

+++ ewm-service/src/main/java/ru/practicum/ewm/category/service/CategoryServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.category.dto.CategoryDto;
 import ru.practicum.ewm.category.dto.CategoryDtoOut;
+import ru.practicum.ewm.event.repository.EventRepository;
 import ru.practicum.ewm.exception.NotFoundException;
 import ru.practicum.ewm.category.mapper.CategoryMapper;
 import ru.practicum.ewm.category.model.Category;
@@ -18,6 +19,7 @@ import java.util.Collection;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    public final EventRepository eventRepository;
 
     @Override
     public Collection<CategoryDtoOut> getAll(Integer offset, Integer limit) {
@@ -57,6 +59,14 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new NotFoundException("Category", id);
+        }
+
+        if (eventRepository.existsByCategoryId(id)) {
+            throw new IllegalStateException("Cannot delete category. There are events associated with it.");
+        }
+
         categoryRepository.deleteById(id);
     }
 }
